GetMapping : 단건조회

+++ src/main/java/repository/MemberRepository.java
@@ -3,5 +3,5 @@ package repository;
 import entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 }
